refactor(acerca): modularize about page
