+++ src/features/acerca/components/Hero.jsx
@@ -1,0 +1,16 @@
+/**
+ * Hero section for the Acerca page.
+ */
+import React from 'react';
+import styles from '../../../styles/acerca.module.css';
+
+const Hero = () => (
+  <section className={styles.hero}>
+    <div className={styles.heroContent}>
+      <h1>Sobre App_Rural</h1>
+      <p>Transparencia y monitoreo para sistemas de agua potable rurales.</p>
+    </div>
+  </section>
+);
+
+export default Hero;
